Updated OsDriver

+++ src/edu/kennesaw/core/memory/Memory.java
@@ -1,6 +1,6 @@
 package edu.kennesaw.core.memory;
 
 public interface Memory {
-    void write(int address, int[] data) throws MemoryOverflowException;
+    void write(int address, byte[] data) throws MemoryOverflowException;
     byte[] read(int address, int length) throws MemoryOverflowException;
 }
